pyproject.toml
[project]
name = "morning-byte"
version = "0.1.0"
description = "Daily tech news digest delivered as beautiful EPUBs to your Kindle"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "python-dateutil>=2.8.0",
    "feedparser>=6.0.0",
    "markdownify>=0.11.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
morning-byte = "morning_byte.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/morning_byte"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
